sched/fair: fix prefer_idle behaviour
prefer_idle is meant for reducing wake_latencies for the task and
hence such a task needs to be placed on an idle cpu and if idle
isn't available, the most spare active cpu.

Currently the code limits the search to the cluster of the
start_cpu, i.e. if an active cpu is found it uses it and doesn't
check for more clusters for idle cpu.

Fix this by allowing to search more clusters regardless of a cpu
is found. But still keep the optimization to break from searching
more groups based on schedtune boost and the capacity of next
cluster.

Change-Id: Ia488aabab448af83e427f6f57da8a7baae40278d
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>
Signed-off-by: MadeOfGreat <ravenklawasd@gmail.com>
Signed-off-by: KangMamles <imzea69@gmail.com>